:construction: Attempt at passing build
This is a WIP, a PR will only be opened to trgger a Travis build.

The Python 2.6 build erros on OrderedDict - not trying to fix that yet.

The `depart_row` in latex_mods appears to be simplified upstream, may
not need to overide the method here any longer.

In `visit_figure`, the sets for hyperlinks have changed to a dict keyed
on the set type, and popping the single key from the dict accomplishes
the clearing aspect.